code-templates/api-integrations/google-ads-api/requirements.txt
google-ads==23.0.0
